+++ ChangeLog.md
@@ -2,11 +2,15 @@
 
 ## Versions ##
 
-### Unreleased
+### 2.10.0 (2020-04-01)
 
 * Improvements
   * Adds 'IgnoreHardwareCheckFailure' property to XD7Feature and XD7Features resources (CVAD 1906+)
   * Adds 'FAS' role to XD7Feature and XD7Features resources
+  * Adds 'MinimumFunctionalLevel' property to XD7Catalog
+
+* Bug Fixes
+  * Adds PsDscRunAsCredential support to XD7SiteConfig
 
 ### 2.9.0 (2020-03-02)